Move stream tests to their own file
